requirements.txt
streamlit
pandas
torch
torchvision
transformers
Pillow
numpy
scikit-learn
